Moving datalad code to it's own module
Datalad code was previously placed under common, but now that the time update will be in it's own file, it's better to place these functions in another file, hence, for clarity purposes, it's better to create an independent module.
